requirements.txt
flask
flask-cors
python-dotenv
langchain-openai 
psycopg2-binary
faker 
mcp_use